Enable simplecov and make files more readable (#4)
* Make files more readable

The comment cruft is ultimately just painful and I now always wind up
removing it.  This gives you the option to do it when the repo is
commissioned.

* Enable simplecov with 100 coverage required

This is the standard for all the projects I've worked on recently.
Adjust according to your needs...

* Keep SimpleCov from breaking

Calling SimpleCov.start with a config block breaks on this version. The
exclusions being passed in the block no longer seem to be required, so
these have been removed.